+++ src/images.ts
@@ -1,11 +1,12 @@
-import { JpegEmbedder, PDFDocument, PDFRef, toUint8Array } from 'pdf-lib';
+import { JpegEmbedder, PDFDocument, PDFRef, PngEmbedder, toUint8Array } from 'pdf-lib';
 
 import { parseBinaryData } from './binary-data.js';
-import { readAs, readObject, required } from './types.js';
+import { optional, readAs, readObject, required, types } from './types.js';
 
 export type ImageDef = {
   name: string;
   data: string | Uint8Array | ArrayBuffer;
+  format: 'jpeg' | 'png';
 };
 
 export type Image = {
@@ -13,26 +14,33 @@ export type Image = {
   width: number;
   height: number;
   data: Uint8Array;
+  format: 'jpeg' | 'png';
   pdfRef?: PDFRef;
 };
 
 export function readImages(input: unknown): ImageDef[] {
   return Object.entries(readObject(input)).map(([name, imageDef]) => {
-    const { data } = readAs(imageDef, name, required(readImage));
-    return { name, data };
+    const { data, format } = readAs(imageDef, name, required(readImage));
+    return { name, data, format: format ?? 'jpeg' };
   });
 }
 
-function readImage(input: unknown): { data: Uint8Array } {
-  return readObject(input, { data: required(parseBinaryData) }) as { data: Uint8Array };
+function readImage(input: unknown) {
+  return readObject(input, {
+    data: required(parseBinaryData),
+    format: optional(types.string({ enum: ['jpeg', 'png'] })),
+  }) as { data: Uint8Array; format?: 'jpeg' | 'png' };
 }
 
 export async function loadImages(imageDefs: ImageDef[]): Promise<Image[]> {
   return await Promise.all(
     imageDefs.map(async (def) => {
       const data = toUint8Array(def.data);
-      const { width, height } = await JpegEmbedder.for(data);
-      return { name: def.name, width, height, data };
+      const embedder = await (def.format === 'png'
+        ? PngEmbedder.for(data)
+        : JpegEmbedder.for(data));
+      const { width, height } = embedder;
+      return { name: def.name, format: def.format, data, width, height };
     })
   );
 }
@@ -42,7 +50,9 @@ export function registerImage(image: Image, pdfDoc: PDFDocument) {
   (pdfDoc as any).images.push({
     async embed() {
       try {
-        const embedder = await JpegEmbedder.for(image.data);
+        const embedder = await (image.format === 'png'
+          ? PngEmbedder.for(image.data)
+          : JpegEmbedder.for(image.data));
         embedder.embedIntoContext(pdfDoc.context, ref);
       } catch (error) {
         throw new Error(